feature:1.将long类型的数据都替换成了Long；2.修复了一些小bug 3.新增SpringSecurity模块，准备开始重构认证授权框架

+++ common/service-util/src/main/java/com/feidian/util/JwtUtil.java
@@ -93,7 +93,7 @@ public class JwtUtil {
             Jws<Claims> claimsJws = Jwts.parser().setSigningKey(generalKey()).parseClaimsJws(token);
             Claims claims = claimsJws.getBody();
             String userId = (String) claims.get("sub");
-            return Integer.parseInt(userId);
+            return Long.parseLong(userId);
 
         } catch (Exception e) {
             e.printStackTrace();
